refactor(native): colocate move-target modules under runtime input
