+++ src/native/engine/runtime/input/input_abi.h
@@ -3,7 +3,7 @@
 
 #include "../engine/engine_composition.h"
 #include "../engine/engine_state.h"
-#include "../move_target_service.h"
+#include "move_target_service.h"
 
 #ifdef __cplusplus
 extern "C"
