+++ tests/native/architecture/runtime_architecture_guard_rules.h
@@ -47,7 +47,7 @@ static const FileRule kRuntimeArchitectureGuardRules[] = {
         1,
     },
     {
-        "runtime/move_target_service.c",
+        "runtime/input/move_target_service.c",
         "domain services must not depend on orchestration",
         {"orchestration/"},
         1,
